added people service and uaa service
