+++ build.gradle
@@ -1,80 +1,1 @@
-buildscript {
-    ext.spek_version="2.0.2"
-    ext.coroutines_version="1.2.0"
-    ext.kotlin_version = '1.3.30'
-    ext.junit_version = '5.2.0'
-}
-
-plugins {
-    id 'org.springframework.boot' version '2.1.3.RELEASE'
-    id 'org.jetbrains.kotlin.jvm' version '1.3.30'
-    id 'org.jetbrains.kotlin.plugin.spring' version '1.3.30'
-    id 'com.google.cloud.tools.jib' version '1.0.1'
-}
-
-apply plugin: 'io.spring.dependency-management'
-
-group = 'zhi.yest.vk'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '1.8'
-
-repositories {
-    mavenCentral()
-    mavenLocal()
-    maven { url "http://dl.bintray.com/spekframework/spek" }
-}
-
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-webflux'
-    implementation 'org.springframework.boot:spring-boot-starter-security'
-    implementation 'org.springframework.security:spring-security-oauth2-client'
-    compileOnly group: 'javax.servlet', name: 'javax.servlet-api', version: '4.0.1'
-
-    implementation 'com.fasterxml.jackson.module:jackson-module-kotlin'
-    implementation 'org.jetbrains.kotlin:kotlin-reflect'
-    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8'
-    implementation group: 'org.jetbrains.kotlinx', name: 'kotlinx-coroutines-core', version: "$coroutines_version"
-    implementation group: 'org.jetbrains.kotlinx', name: 'kotlinx-coroutines-reactive', version: "$coroutines_version"
-    implementation 'zhi.yest:vk-method-executor:1.0.2'
-    runtimeOnly 'org.springframework.boot:spring-boot-devtools'
-
-    testImplementation "com.nhaarman.mockitokotlin2:mockito-kotlin:2.+"
-    testRuntimeOnly "org.junit.jupiter:junit-jupiter-engine:$junit_version"
-    testImplementation "org.junit.jupiter:junit-jupiter-api:$junit_version"
-    testImplementation "org.spekframework.spek2:spek-dsl-jvm:$spek_version"
-    testRuntimeOnly "org.spekframework.spek2:spek-runner-junit5:$spek_version"
-    
-    annotationProcessor "org.springframework.boot:spring-boot-configuration-processor"
-}
-
-test {
-    useJUnitPlatform {
-        includeEngines 'spek2'
-    }
-}
-
-compileKotlin {
-    kotlinOptions {
-        freeCompilerArgs = ['-Xjsr305=strict']
-        jvmTarget = '1.8'
-    }
-}
-
-compileTestKotlin {
-    kotlinOptions {
-        freeCompilerArgs = ['-Xjsr305=strict']
-        jvmTarget = '1.8'
-    }
-}
-
-jib {
-    to {
-        image = 'dazzbourgh/friend-finder'
-        tags = ['latest']
-    }
-    container {
-        jvmFlags = ['-XX:+UnlockExperimentalVMOptions',
-                    '-XX:+UseCGroupMemoryLimitForHeap',
-                    '-XX:MaxRAMFraction=1']
-    }
-}
+//TODO
