+++ people-service/src/main/kotlin/zhi/yest/communityintersection/peopleservice/domain/Request.kt
@@ -1,0 +1,4 @@
+package zhi.yest.communityintersection.peopleservice.domain
+
+data class Request(val groupIds: List<String>,
+                   val fields: Map<String, String>)
