+++ people-service/src/main/kotlin/zhi/yest/communityintersection/peopleservice/util/UtilFunctions.kt
@@ -1,0 +1,5 @@
+package zhi.yest.communityintersection.peopleservice.util
+
+fun String.trimQuotes(): String {
+    return this.replace("\"", "")
+}
