+++ people-service/src/main/kotlin/zhi/yest/communityintersection/peopleservice/vk/DelayingRequestSender.kt
@@ -1,0 +1,5 @@
+package zhi.yest.communityintersection.peopleservice.vk
+
+interface DelayingRequestSender {
+    suspend fun <T> request(block: suspend () -> T): T
+}
