--- src/main/kotlin/zhi/yest/vk/friendfinder/domain/Request.kt
@@ -1,4 +1,0 @@
-package zhi.yest.vk.friendfinder.domain
-
-data class Request(val groupIds: List<String>,
-                   val fields: Map<String, String>)
