--- src/main/kotlin/zhi/yest/vk/friendfinder/util/UtilFunctions.kt
@@ -1,5 +1,0 @@
-package zhi.yest.vk.friendfinder.util
-
-fun String.trimQuotes(): String {
-    return this.replace("\"", "")
-}
