--- src/main/kotlin/zhi/yest/vk/friendfinder/vk/DelayingRequestSender.kt
@@ -1,5 +1,0 @@
-package zhi.yest.vk.friendfinder.vk
-
-interface DelayingRequestSender {
-    suspend fun <T> request(block: suspend () -> T): T
-}
