--- src/main/resources/application-local.yml
@@ -1,1 +1,0 @@
-base-url: http://localhost:8080
