--- src/main/resources/application.yml
@@ -1,23 +1,0 @@
-server:
-  port: ${PORT:8080}
-vk:
-  request:
-    interval: 500
-  api:
-    version: 5.95
-spring:
-  security:
-    oauth2:
-      client:
-        registration:
-          vk:
-            client-id: ${client_id}
-            client-secret: ${client_secret}
-            authorization-grant-type: authorization_code
-            scopes: read
-            redirect-uri: ${base-url}/login/oauth2/code/vk
-        provider:
-          vk:
-            authorization-uri: https://oauth.vk.com/authorize
-            token-uri: https://oauth.vk.com/access_token
-            user-info-uri: https://api.vk.com/method/users.get
